updated files within relocated files
