+++ client/src/components/CreateProjectContainer.js
@@ -2,7 +2,7 @@
 
 import { connect } from 'react-redux';
 import { createProject } from '../../actions';
-import AddProject from '../AddProject';
+import AddProject from './AddProject';
 
 const mapDispatchToProps = dispatch => {
   return {
