+++ client/src/components/ViewAProfileProjectContainer.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import ViewProfileListing from '../ViewProfileListing.js';
-import { deleteProject, viewProjects } from '../../actions';
+import ViewProfileListing from './ViewProfileListing.js';
+import { deleteProject, viewProjects } from '../actions';
 import { } from "react-router-dom";
-import Profile from '../Profile';
+import Profile from './Profile';
 
 function ViewAProfileProject({ projects, onView }) {
