+++ client/src/components/ViewAProjectContainer.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import ViewListing from '../ViewListing.js';
-import { deleteProject, viewProjects } from '../../actions';
+import ViewListing from './ViewListing.js';
+import { deleteProject, viewProjects } from '../actions';
 import { } from "react-router-dom";
-import Projects from '../ProjectsRedux.js';
+import Projects from './ProjectsRedux.js';
 
 function ViewAProject({ projects, onView }) {
